We need a way to combine multiple embed maps together at runtime.
Object.defineProperty will fail if the property already exists.

So instead of using defineProperty like before, we define the property as an array and push embed maps onto it.

This way, at runtime we will have an array of size 1 or greater and in the output bundle an array of size 1 (merge + hoist).